src/requirements.txt
arrow==1.2.2
asgiref==3.5.0
attrs==21.4.0
binaryornot==0.4.4
black==22.1.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
cookiecutter==1.7.3
crispy-tailwind==0.5.0
cryptography==36.0.1
defusedxml==0.7.1
Django==3.2.12
django-allauth==0.49.0
django-browser-reload==1.3.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-tailwind==3.1.1
flake8==4.0.1
idna==3.3
iniconfig==1.1.1
Jinja2==3.0.3
jinja2-time==0.2.0
MarkupSafe==2.1.0
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.2.0
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.1
pluggy==1.0.0
poyo==0.5.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.7
pytest==7.0.1
pytest-django==4.5.2
python-dateutil==2.8.2
python-slugify==6.1.0
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.8
